+++ lib/index.ts
@@ -1,3 +1,4 @@
 export { Server, type ServerOptions } from "./server";
+export { Socket, type CloseReason } from "./socket";
 export { type RawData } from "./parser";
 export { type BunWebSocket, type WebSocketData } from "./transports/websocket";